### Description
Currently OSB prints out an error when users invoke `opensearch-benchmark` only. This PR enables OSB to display the help message when no arguments are passed. 
Before suggested change:
```
$ opensearch-benchmark
Traceback (most recent call last):
  File "/home/ec2-user/.local/bin/opensearch-benchmark", line 8, in <module>
    sys.exit(main())
  File "/home/ec2-user/.local/lib/python3.9/site-packages/osbenchmark/benchmark.py", line 949, in main
    console.init(quiet=args.quiet)
AttributeError: 'Namespace' object has no attribute 'quiet'
```
After suggested change:
```
hoangia@3c22fbd0d988 opensearch-benchmark % opensearch-benchmark
usage: opensearch-benchmark [-h] [--version]

   ____                  _____                      __       ____                  __                         __
  / __ \____  ___  ____ / ___/___  ____ ___________/ /_     / __ )___  ____  _____/ /_  ____ ___  ____ ______/ /__
 / / / / __ \/ _ \/ __ \\__ \/ _ \/ __ `/ ___/ ___/ __ \   / __  / _ \/ __ \/ ___/ __ \/ __ `__ \/ __ `/ ___/ //_/
/ /_/ / /_/ /  __/ / / /__/ /  __/ /_/ / /  / /__/ / / /  / /_/ /  __/ / / / /__/ / / / / / / / / /_/ / /  / ,<
\____/ .___/\___/_/ /_/____/\___/\__,_/_/   \___/_/ /_/  /_____/\___/_/ /_/\___/_/ /_/_/ /_/ /_/\__,_/_/  /_/|_|
    /_/

 A benchmarking tool for OpenSearch

options:
  -h, --help  show this help message and exit
  --version   show program's version number and exit

Find out more about Benchmark at https://opensearch.org/docs
```

### Issues Resolved
#237 

### Testing
- [x] New functionality includes testing

- Ran Unittests 
- Ran opensearch-benchmark test in test-mode
- Ran basic test execution with OpenSearch 1.0.0 distribution

Basic test execution with OpenSearch 1.0.0 distribution 
```
hoangia@3c22fbd0d988 opensearch-benchmark % opensearch-benchmark execute_test --distribution-version=1.0.0 --workload=geonames --test-mode


   ____                  _____                      __       ____                  __                         __
  / __ \____  ___  ____ / ___/___  ____ ___________/ /_     / __ )___  ____  _____/ /_  ____ ___  ____ ______/ /__
 / / / / __ \/ _ \/ __ \\__ \/ _ \/ __ `/ ___/ ___/ __ \   / __  / _ \/ __ \/ ___/ __ \/ __ `__ \/ __ `/ ___/ //_/
/ /_/ / /_/ /  __/ / / /__/ /  __/ /_/ / /  / /__/ / / /  / /_/ /  __/ / / / /__/ / / / / / / / / /_/ / /  / ,<
\____/ .___/\___/_/ /_/____/\___/\__,_/_/   \___/_/ /_/  /_____/\___/_/ /_/\___/_/ /_/_/ /_/ /_/\__,_/_/  /_/|_|
    /_/

[INFO] Preparing for test execution ...
[INFO] Downloading OpenSearch 1.0.0 (411.1 MB total size)                           [100%]
[INFO] Downloading workload data (20.5 kB total size)                             [100.0%]
[INFO] Decompressing workload data from [/Users/hoangia/.benchmark/benchmarks/data/geonames/documents-2-1k.json.bz2] to [/Users/hoangia/.benchmark/benchmarks/data/geonames/documents-2-1k.json] ... [OK]
[INFO] Preparing file offset table for [/Users/hoangia/.benchmark/benchmarks/data/geonames/documents-2-1k.json] ... [OK]
[INFO] Executing test with workload [geonames], test_procedure [append-no-conflicts] and provision_config_instance ['defaults'] with version [1.0.0].

Running delete-index                                                           [100% done]
Running create-index                                                           [100% done]
Running check-cluster-health                                                   [100% done]
Running index-append                                                           [100% done]
Running refresh-after-index                                                    [100% done]
Running force-merge                                                            [100% done]
Running refresh-after-force-merge                                              [100% done]
Running wait-until-merges-finish                                               [100% done]
Running index-stats                                                            [100% done]
Running node-stats                                                             [100% done]
Running default                                                                [100% done]
Running term                                                                   [100% done]
Running phrase                                                                 [100% done]
Running country_agg_uncached                                                   [100% done]
Running country_agg_cached                                                     [100% done]
Running scroll                                                                 [100% done]
Running expression                                                             [100% done]
Running painless_static                                                        [100% done]
Running painless_dynamic                                                       [100% done]
Running decay_geo_gauss_function_score                                         [100% done]
Running decay_geo_gauss_script_score                                           [100% done]
Running field_value_function_score                                             [100% done]
Running field_value_script_score                                               [100% done]
Running large_terms                                                            [100% done]
Running large_filtered_terms                                                   [100% done]
Running large_prohibited_terms                                                 [100% done]
Running desc_sort_population                                                   [100% done]
Running asc_sort_population                                                    [100% done]
Running asc_sort_with_after_population                                         [100% done]
Running desc_sort_geonameid                                                    [100% done]
Running desc_sort_with_after_geonameid                                         [100% done]
Running asc_sort_geonameid                                                     [100% done]
Running asc_sort_with_after_geonameid                                          [100% done]

------------------------------------------------------
    _______             __   _____
   / ____(_)___  ____ _/ /  / ___/_________  ________
  / /_  / / __ \/ __ `/ /   \__ \/ ___/ __ \/ ___/ _ \
 / __/ / / / / / /_/ / /   ___/ / /__/ /_/ / /  /  __/
/_/   /_/_/ /_/\__,_/_/   /____/\___/\____/_/   \___/
------------------------------------------------------

|                                                         Metric |                           Task |       Value |    Unit |
|---------------------------------------------------------------:|-------------------------------:|------------:|--------:|
|                     Cumulative indexing time of primary shards |                                |      0.0532 |     min |
|             Min cumulative indexing time across primary shards |                                |  0.00548333 |     min |
|          Median cumulative indexing time across primary shards |                                |  0.00656667 |     min |
|             Max cumulative indexing time across primary shards |                                |   0.0264333 |     min |
|            Cumulative indexing throttle time of primary shards |                                |           0 |     min |
|    Min cumulative indexing throttle time across primary shards |                                |           0 |     min |
| Median cumulative indexing throttle time across primary shards |                                |           0 |     min |
|    Max cumulative indexing throttle time across primary shards |                                |           0 |     min |
|                        Cumulative merge time of primary shards |                                |           0 |     min |
|                       Cumulative merge count of primary shards |                                |           0 |         |
|                Min cumulative merge time across primary shards |                                |           0 |     min |
|             Median cumulative merge time across primary shards |                                |           0 |     min |
|                Max cumulative merge time across primary shards |                                |           0 |     min |
|               Cumulative merge throttle time of primary shards |                                |           0 |     min |
|       Min cumulative merge throttle time across primary shards |                                |           0 |     min |
|    Median cumulative merge throttle time across primary shards |                                |           0 |     min |
|       Max cumulative merge throttle time across primary shards |                                |           0 |     min |
|                      Cumulative refresh time of primary shards |                                |   0.0221667 |     min |
|                     Cumulative refresh count of primary shards |                                |          30 |         |
|              Min cumulative refresh time across primary shards |                                |     0.00285 |     min |
|           Median cumulative refresh time across primary shards |                                |      0.0034 |     min |
|              Max cumulative refresh time across primary shards |                                |     0.00895 |     min |
|                        Cumulative flush time of primary shards |                                |           0 |     min |
|                       Cumulative flush count of primary shards |                                |           0 |         |
|                Min cumulative flush time across primary shards |                                |           0 |     min |
|             Median cumulative flush time across primary shards |                                |           0 |     min |
|                Max cumulative flush time across primary shards |                                |           0 |     min |
|                                        Total Young Gen GC time |                                |       0.018 |       s |
|                                       Total Young Gen GC count |                                |           2 |         |
|                                          Total Old Gen GC time |                                |           0 |       s |
|                                         Total Old Gen GC count |                                |           0 |         |
|                                                     Store size |                                | 0.000557473 |      GB |
|                                                  Translog size |                                | 2.56114e-07 |      GB |
|                                         Heap used for segments |                                |    0.158394 |      MB |
|                                       Heap used for doc values |                                |   0.0175705 |      MB |
|                                            Heap used for terms |                                |    0.113525 |      MB |
|                                            Heap used for norms |                                |   0.0151978 |      MB |
|                                           Heap used for points |                                |           0 |      MB |
|                                    Heap used for stored fields |                                |   0.0121002 |      MB |
|                                                  Segment count |                                |          26 |         |
|                                                 Min Throughput |                   index-append |      1686.2 |  docs/s |
|                                                Mean Throughput |                   index-append |      1686.2 |  docs/s |
|                                              Median Throughput |                   index-append |      1686.2 |  docs/s |
|                                                 Max Throughput |                   index-append |      1686.2 |  docs/s |
|                                        50th percentile latency |                   index-append |     533.637 |      ms |
|                                       100th percentile latency |                   index-append |     804.877 |      ms |
|                                   50th percentile service time |                   index-append |     533.637 |      ms |
|                                  100th percentile service time |                   index-append |     804.877 |      ms |
|                                                     error rate |                   index-append |           0 |       % |
|                                                 Min Throughput |       wait-until-merges-finish |       39.07 |   ops/s |
|                                                Mean Throughput |       wait-until-merges-finish |       39.07 |   ops/s |
|                                              Median Throughput |       wait-until-merges-finish |       39.07 |   ops/s |
|                                                 Max Throughput |       wait-until-merges-finish |       39.07 |   ops/s |
|                                       100th percentile latency |       wait-until-merges-finish |     25.0307 |      ms |
|                                  100th percentile service time |       wait-until-merges-finish |     25.0307 |      ms |
|                                                     error rate |       wait-until-merges-finish |           0 |       % |
|                                                 Min Throughput |                    index-stats |       88.42 |   ops/s |
|                                                Mean Throughput |                    index-stats |       88.42 |   ops/s |
|                                              Median Throughput |                    index-stats |       88.42 |   ops/s |
|                                                 Max Throughput |                    index-stats |       88.42 |   ops/s |
|                                       100th percentile latency |                    index-stats |     19.2001 |      ms |
|                                  100th percentile service time |                    index-stats |     7.49804 |      ms |
|                                                     error rate |                    index-stats |           0 |       % |
|                                                 Min Throughput |                     node-stats |       85.32 |   ops/s |
|                                                Mean Throughput |                     node-stats |       85.32 |   ops/s |
|                                              Median Throughput |                     node-stats |       85.32 |   ops/s |
|                                                 Max Throughput |                     node-stats |       85.32 |   ops/s |
|                                       100th percentile latency |                     node-stats |     18.7108 |      ms |
|                                  100th percentile service time |                     node-stats |     6.64253 |      ms |
|                                                     error rate |                     node-stats |           0 |       % |
|                                                 Min Throughput |                        default |       12.49 |   ops/s |
|                                                Mean Throughput |                        default |       12.49 |   ops/s |
|                                              Median Throughput |                        default |       12.49 |   ops/s |
|                                                 Max Throughput |                        default |       12.49 |   ops/s |
|                                       100th percentile latency |                        default |     90.5016 |      ms |
|                                  100th percentile service time |                        default |     10.1879 |      ms |
|                                                     error rate |                        default |           0 |       % |
|                                                 Min Throughput |                           term |        78.3 |   ops/s |
|                                                Mean Throughput |                           term |        78.3 |   ops/s |
|                                              Median Throughput |                           term |        78.3 |   ops/s |
|                                                 Max Throughput |                           term |        78.3 |   ops/s |
|                                       100th percentile latency |                           term |     17.4888 |      ms |
|                                  100th percentile service time |                           term |      4.5026 |      ms |
|                                                     error rate |                           term |           0 |       % |
|                                                 Min Throughput |                         phrase |       44.39 |   ops/s |
|                                                Mean Throughput |                         phrase |       44.39 |   ops/s |
|                                              Median Throughput |                         phrase |       44.39 |   ops/s |
|                                                 Max Throughput |                         phrase |       44.39 |   ops/s |
|                                       100th percentile latency |                         phrase |     28.5227 |      ms |
|                                  100th percentile service time |                         phrase |     5.77905 |      ms |
|                                                     error rate |                         phrase |           0 |       % |
|                                                 Min Throughput |           country_agg_uncached |       15.02 |   ops/s |
|                                                Mean Throughput |           country_agg_uncached |       15.02 |   ops/s |
|                                              Median Throughput |           country_agg_uncached |       15.02 |   ops/s |
|                                                 Max Throughput |           country_agg_uncached |       15.02 |   ops/s |
|                                       100th percentile latency |           country_agg_uncached |     72.6533 |      ms |
|                                  100th percentile service time |           country_agg_uncached |     5.87177 |      ms |
|                                                     error rate |           country_agg_uncached |           0 |       % |
|                                                 Min Throughput |             country_agg_cached |       67.73 |   ops/s |
|                                                Mean Throughput |             country_agg_cached |       67.73 |   ops/s |
|                                              Median Throughput |             country_agg_cached |       67.73 |   ops/s |
|                                                 Max Throughput |             country_agg_cached |       67.73 |   ops/s |
|                                       100th percentile latency |             country_agg_cached |     34.9918 |      ms |
|                                  100th percentile service time |             country_agg_cached |     19.9631 |      ms |
…